Colocate single-use utility functions in NetGraph
